requirements.txt
Flask==3.0.0
flask-redis==0.4.0
redis==5.0.1
rq==1.15.1
twilio==8.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
coverage==7.3.0
flake8==6.1.0
